Combine Home Manager environment (#2)
* Attempt to merge HM config into the flake

Building the NixOS config currently works, but the HM configuration is
trying to import a channel somewhere and I need to find that to get it
to work.

* Get HM working

It works now but it comes at the cost of removing Slack and Discord for
the time being. Just need to figure out how to allow unfree for those to
work.

* Get unfree packages working

Just had to add a line to allow them. It was previously read out of the
home directory which is why it worked before switching to a flake.

* Add chrome-dark Nixpkgs overlay

* Change `$TERM` to xterm in Alacritty

* Switch Chromium to use dark package

* Add Zoom and Chrome